third_party/blink/web_tests/editing/selection/user-select-none-in-editable-expected.txt

PASS selection.focusNode is sample2.firstChild
PASS selection.focusOffset is 2
PASS successfullyParsed is true

TEST COMPLETE
